Added petscreldefs.h to allow PFLOTRAN to be built against PETSc 2.3.2.
Note that petscreldefs.h requires the PETSC_VERSION_RELEASE macro to be
defined as 1.  This is defined if petsc.h is included and Satish's
PETSc 2.3.2 patch to provide PETSC_VERSION_RELEASE is included.  If
building without this patch, you will need to define this macro
when compiling pflowgrid_mod.F90 via a command line option such as
'-DPETSC_VERSION_RELEASE=1'.